Moved some variables out of newReport reducer
isUnsaved and isSavingRequested are now moved to their own reducer. This is so that these values don't get saved inside the reports along with actual report data
